pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.81.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pyrebase>=3.0.18",
    "pyrebase4>=4.8.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
